tpic/requirements.txt
playwright==1.58.0
python-dotenv==1.0.1
colorama==0.4.6
aiohttp==3.9.1
anthropic==0.18.0
google-generativeai==0.3.0
groq==0.4.2
requests==2.31.0
pillow==10.1.0
